Add StockPriceService happy path
